Unfinished refactoring
