+++ book_logic.rb
@@ -1,0 +1,7 @@
+class BookLogic
+  def initialize(list_handler)
+    @list_handler = list_handler
+  end
+
+  
+end
