+++ list_middleware.rb
@@ -1,0 +1,8 @@
+class ListClass
+  def initialize(books_handler, rentals_handler, people_handler)
+    @books = []
+    @rentals = []
+    @people = []
+  end
+  attr_accessor :people, :books, :rentals
+end
